requirements.txt
Flask
flask-cors
uvicorn
tensorflow==2.13.0
pillow
numpy
torch
firebase-admin
torchvision
openv
opencv-python
scipy
natsort
yolov5
scikit-image
scikit-learn
pandas
joblib
textblob
